fix(sdk-ui-dashboard): avoid redux use in DashboardLayout
This would make the DashboardLayout unusable outside of Dashboard
component context, which we need to do for legacy edit mode.

JIRA: RAIL-4437
